Added jwt token

+++ api/server/app/schemas/order.py
@@ -2,7 +2,7 @@ from typing import List, Any , Optional, Union, ClassVar
 from pydantic import BaseModel, validator, Field
 
 class OrderBase(BaseModel):
-    amount : int = Field(..., description="Amount", example=100.00, gt=0) #validation for price >0
+    amount : float = Field(..., description="Amount", example=100.00, gt=0) #validation for price >0
     currency : str = Field(..., description="Currency", example="USD")
     order_id : str = Field(..., description="Order ID", example="order_123")
 
@@ -11,7 +11,7 @@ class OrderCreate(BaseModel):
     examination_ids : List[int]
 
 class OrderAmount(BaseModel):
-    amount : int
+    amount : float
     notes : str
 
 class OrderCapture(BaseModel):
